backends/libinput: Fixed Device::setOutputName not changing the output
Commit f860be3a intoduced a bug where changing target display for
touchscreen or tablet once would prevent it being changed again.
It is caused by setOutputName method updating the output only if
m_outputUuid is not already set.

This change makes both setOutputName and setOutputUuid update stored
name UUID and pointer to output screen and update "OutputUuid" property
in the config file.

setConfigOutputName will change the output only if it is not already
set. It is only used for loading from config and needed for backwards
compatibility with config file created before matching display by UUID
was introduced.

BUG: 510505