OutputDeviceV2Interface: guard for global removed in bind
When we plug in a monitor, we register a global kde_output_device_v2 to
clients. If we unplug the monitor and delete the global while a client
is trying to bind it, we would send a client a global_remove message
*before* messages to the bound object ID. This caused Wayland clients to
not recognize the object IDs the server tried sending to the client. The
next time the server tried sending the client an object ID, it would be
greater than the last object ID the client saw + 1, causing the client
app to exit with errors like:
	not a valid new object id (4278190164), message mode(n)
	The Wayland connection experienced a fatal error: Invalid argument

Fix this bug by not sending messages from objects belonging to globals
we've already removed.

This is safe because both binding (from wl_event_loop_dispatch) and
removing outputs (from KWin::DrmBackend::handleUdevEvent) run in the
same main thread.

BUG: 510734