Revert "Check that output passed to Workspace::setActiveOutput() is valid"
This reverts commit 5225bfed3316cec7ea411e7d72990b312843b453.

While this seemed like a good idea at the time, the Workspace may set an
output, which is not in the m_outputs list, while tearing down layer shell
windows, etc.

It doesn't mean that the Workspace will end up with a dangling active
output. The Workspace can still recover, crashing in that case will be
far from ideal.