Fix Workspace::activateNextWindow() corrupting active output on hotplug
After 4ee32d7f71de7f7f9a24c44b9dc4dd31fd3be98f, layer shell windows are
removed when the Workspace::outputRemoved() signal is emitted.

The side-effect of that change is that the activateNextWindow() may be
called when the Workspace::outputRemoved() signal is emitted.

The activateNextWindow() function is interesting because it can set the
active output to the one of the Window.

If the next candidate window is on a just removed output, the Workspace
can end up with a dangling active output.

To prevent that case, this change moves active output cleanup code so it
is after the outputRemoved() signal is emitted and after windows are
re-arranged.

BUG: 495702