autotests: add a test helper for creating and presenting windows
This helper reduces the boilerplate code needed to write tests, and should make
them more reliable by waiting for each window to be closed before moving on to
the next test case (which has caused problems before and is easy to forget).
Later on, the helper can be expanded more to cover other more common things as
well, like reacting to configure events.