backends/drm: more consistently use the next output state in config tests
Using the "current" output state is problematic when we're testing for a new
output configuration - we're changing the output state and this should check
whether or not that new state actually works.
To ensure we don't have to keep the next state around and in sync with the
current state all the time, this just makes the next state an optional, and
only uses it if we're actually testing an output configuration.

CCBUG: 509635