inputmethod: Unset panel on close
When the InputPanel is deleted (in kwin terms) we shouldn't use it any
more from the InputMethod. In particular the case where we handle a new
panel being created.

The existing use of QPointer doesn't suffice as the object can be kept
in a deleted state for a while for purposes like animations.