plugins/blur: Disable background contrast effect
The background contrast effect was introduced to improve readability
with translucent surfaces. However, after changing the algorithm of the
blur effect and also tweaking its defaults, the effect of the background
contrast pass is negligible. Neither text readability nor Plasma
apperance is severely affected by the omission of the contrast effect.

In 6.5, the background contrast effect and the blur effect got merged to
simplify the overall design and corresponding protocol bits, and to
improve performance.

Unfortunately, the background contrast has created issues too. The most
noticeable is window flashing when animating the opacity of a window.
We found a workaround to minimize it, but it's still there, it depends
how close you look at animated windows and background behind them. It
also made artifacts around rounded panel corners more stronger, although
to be fair, the blur effect has the same issue but it's less noticeable.

The background contrast effect has its own use cases. For example, with
very low blur radii and very translucent windows (or windows with
completely transparent background). However, those are extreme cases,
and beyond what we are willing to support.

This change introduces a hidden UseContrast option and disables the
background contrast effect. The background contrast effect is deprecated
and will be removed in a future release.

BUG: 510818