opengl/glshadermanager: add the ability to generate shaders for color pipelines
This increases the flexibility of how we can do color conversions, any changes in the pipeline
are automatically mirrored in the shader. More importantly, we will be able to replace the icc
shader with it, and allow effects to set custom color pipelines on windows and outputs.

As some 3rd party effects still rely on it, the infrastructure for the previous
colormanagement.glsl include is left as-is. I don't have any concrete plans on how to expose
this new API to effects that use custom shaders, but it's not too pressing at the moment.