plugins/blur: Use mathematically correct opacity curve for contrast effect
Currently, the contrast effect sets an opacity that is the square of the
opacity of the window in order to disguise itself during fade in/out
animations. However this is incorrect as the resulting animation
brightness graph has a maxima before the end of the animation, causing
a flash of brightness in dark themes and a flash of darkness in light ones.
Instead, we use a curve that's more steep and that, when rendered,
results in a perfectly linear opacity animation.

Note that this formula requres knowing the (dominant) transparency of
the window (e.g. 0.85 for the default Plasma theme). An option would be
to have the application tell the blur effect about its dominant
transparency. However that should not be necessary as setting the goal
transparency to 0.85 has few drawbacks that only appear for almost-opaque
and almost-transparent windows, which do not need the contrast effect
anyway.