xwayland: Manage WlSource using std::unique_ptr
According to 4150c1f047b8ff186d1ff8229a2d79dfe03202e2, the WlSource is
deleted later to avoid sending requests that can confuse klipper, but
nobody listens to the destroyed() signal of the WlSource. The WlSource
also doesn't own any objects that will be destroyed together with it.
Neither the WlSource nor the SelectionSource have defined destructors
that have custom logic too.

Since memory management is one of the problematic areas in the xwayland
code, it's worth taking risk and making the WlSource managed using
std::unique_ptr so it's harder to shoot yourself in the foot when working
with this code.