scene/cursoritem: always reset the surface item if the surface is nullptr
The previous code misbehaved when a client first destroys the cursor surface and
only afterwards sets the cursor surface to null.
This is because SurfaceItemWayland::m_surface is a QPointer, so comparisons with
a nullptr would return true after the surface is destroyed, preventing the surface
item from being deleted.

BUG: 511075