plugins/screenshot,-cast: hide the requesting app's windows in screenshots/casts
This avoids capturing Spectacle's animations, without having to disable animations for it
entirely, and without the potential for races between Spectacle hiding its window and the
screencast actually starting.

This doesn't currently work for apps going through the xdg desktop portal, as the pid of
the client is the portal, and not the actual app originally requesting the screenshot/cast.
This can be fixed with some protocol/API additions, but that can be done later.