wayland/tools: greatly increase global removal timeout
While normally 5 seconds should be plenty, sometimes apps are stupid or buggy and
do blocking things on the main thread, and may crash on output hotplug events
because we delete the global before they react to its removal.
There isn't much of a cost for keeping the global around for a longer time, so this
just extends the timeout to 5 minutes.