xwayland: Simplify selection timestamp tracking
At the moment, there are two levels of timestamps - one in the Selection
and another is in the SelectionSource.

This change simplifies the timestamp business a little so only the Selection
keeps tracks of the selection timestamp.

In long term, it can be used to kill SelectionSource in favor of
AbstractDataSource.