Use an interactive move for toplevel drag
This aligns more with the expecations what happens when a window
is moved by the user. Enables for example  to quick tile the
attached window and enables effects such as wobbly windows to affect it.
Simplifies window tracking a bit, we don't need to care about previous
windows as a new toplevel can only be attached when the previous
toplevel has been unmapped.
BUG: 511570
FIXED-IN: 6.5.3