workspace: add support for automatic brightness adjustments
The brightness setting of each output with auto brightness enabled is adjusted
automatically based on light sensor measurements.
The curve that decides which setting the display should be configured towards
is relatively simple: It's just 6 lux values, each for a 20% brightness step.

When the user adjusts the brightness manually, with either slider or brightness
keys, the curve is adjusted automatically to match. That adjustment is super
basic right now, but it seems to be good enough for the feature to be usable,
and can be improved later on.

BUG: 502122