workspace: add support for automatic brightness adjustments
The brightness setting of each output with auto brightness enabled is adjusted
automatically based on light sensor measurements.
The curve that decides which setting the display should be configured towards
is super simple right now, just a plain linear equation.
When the user adjusts the brightness manually, with either slider or brightness
keys, the "curve" is adjusted automatically to match.
More sophisticated curves or algorithms should be added later on, but this at
least adds the needed infrastructure to do that more easily.

BUG: 502122