opengl/colormanagement: also clip after tonemapping
With night light, the maximum luminance that can be represented in the buffer
without color shifts is lower than the value represented by 1.0. When writing
into the buffer, OpenGL implicitly clips for us to 1.0, so we end up with an
out of bounds value, which ends up causing color shifts when the result is
converted to a different white point.
To avoid that, explicitly clip to the maximum luminance after tonemapping.

BUG: 506212


(cherry picked from commit d70ee5efc9d6b856a067ed284896481cd59a91ac)

Co-authored-by: Xaver Hugl <xaver.hugl@kde.org>