ci: increase test timeout to 90s
Most of testLockScreen's failures seem to be just the test legitimately taking
too much time - every test case waits for lock and unlock, so it just takes a
while. That should ideally be solved in some way too, but I can't think of a
good way to do it right now.
To make sure legit hangs or other issues don't make CI take longer, this reduces
the amount of repeated attempts to 3.