workspace: don't activate or restack popup windows
Activating a popup window implicitly causes its parent to be activated as well,
which makes it get in the way of the previously active window.

Restacking them below the active window sounds sensible on the surface, but also
happens to raise both the popup and its parent above the active window, because
the popup is in a different layer than the active window.
I would like to change restackWindowUnderActive to be more intuitive, but it
looks a bit risky for a bugfix. It can be done later as a refactor.

BUG: 511631