backends/drm: add a timeout to checkCrtcAssignment
When the driver completely rejects the output configuration, KWin ended up freezing
for what looks to the user an indefinite amount of time.

As the chance of an output configuration working becomes more and more slim, the more
atomic tests are done, add a timeout of 3s for each output configuration, which should
in most situations give the user back a responsive system within less than 10 seconds
(3s for normal tests + 3s for low bandwidth fallback + another test for the old config)
and not affect actually working systems in any way.