killprompt: Use unprivileged activation token
There's no real reason this should be more special than any other
dialog prompt. When closing an unresponsive application, the prompt
shouldn't steal focus when you're doing something else in the meantime.

For this, also request the token at the time of the ping rather than
when the prompt actually opens. This captures the user's intent more
closely. While at it, also set the originating surface.