xwayland: Prevent XwlDataSource from unwrapping pipe fds
Signals are not the optimal mechanism for passing the ownership of a
file descriptor. It's unclear who should close it, or whether the file
descriptor is going to be closed at all. The code is the way it is
because back at the time, we had no the FileDescriptor helper.

This change makes the XwlDataSource ask the associated selection to
start a transfer directly. This way, the file descriptor doesn't need to
be unwrapped, and it should be harder to shoot yourself in the foot
while working with this code.