workspace: switch output order configuration to use priority values
Before this commit, output order was simply a list of outputs, with an index
set by kscreen determining the position of the output in that list. This has
some nice advantages, like a guaranteed strict and consistent order between
the outputs.
However, this setup becomes a problem once not every backend output has a
logical output anymore, which will be the case with mirroring.
If the mirroring output doesn't appear in the list, then libkscreen can't know
the (in KWin persistent) priority value of the output, and assumes it's no
longer enabled.
However, if the mirroring output *is* in the list, then plasmashell gets
confused, because as far as it is concerned, this output doesn't exist at all.

To solve that problem, the output device protocol is changed to have a
"priority" event, so libkscreen no longer has to use the output order protocol
to get the priority value of each output, and KWin can just drop outputs from
the output order list, without causing problems in libkscreen.

As a nice side effect, this change also simplifies some code in both KWin and
libkscreen.