workspace: implement X11<->Wayland activation interoperability
It works like this for X11 -> Wayland:
- X11 app posts a "new" message with the app id of the window to be activated
- Wayland client uses the message contents as an activation token

and is implemented by simply setting the activation token to the startup
notification ID (+up to date serial) if the active window belongs to the same
process that posted the "new" message.

For Wayland->X11 it works like this:
- Wayland client creates an activation token
- X11 app posts the "remove" message with the token as the content

and it's implemented by recording the pid of the X11 app and checking it + the
serial of the activation token when activating the window.