xwayland: Drop X11Source
The X11Source type is a thin wrapper around the XwlDataSource. Given how
the code is structured, it is also possible for a Selection to miss some
mime types.

For the async purposes, it's better to send an XConvertSelection()
request for TARGETS and create a data source when the reply arrives.
If there are two pending SelectionNotify events, kwin will not miss
SelectionNotify replies.

Note that there are still races when an X11 and a Wayland client try to
write to the clipboard or the primary selection at the same time.

It will be nice to have a matching setWlSource() setter for X11-owned
data, but I'm not sure how it can be implemented in a less awkward way.
We still need to monitor SeatInterface signals, we still need to claim
or disown the selection when the Wayland data source changes, we still
need to be very careful with the order in which the m_xSource is
destroyed and setSelection() or setPrimarySelection() are called. Not
something critical, but it'll be nice to clean this up more some day.