core: Add Region
The region abstraction is an important component in a compositor. It is
used to determine what area is dirty, what area is occluded, what area
accepts input, and so on.

As we started adding more features on Wayland, we started experiencing
certain limits what QRegion can do and also hitting unexpected API.

This change introduces the Region helper that meets some of our needs,
but it will likely be extended more (and complemented with a RegionF)
to fix fractional scaling support.

The associated logic is not necessarily the easiest so some attention
was paid to tests. In order to test various region operations, all
possible regions for a grid with pre-configured size are generated and
then used for tests. 3x3 was chosen due to the practical reasons, it
covers a pretty good chunk of cases; it'll be great if we could go with
bigger grids, e.g. 4x4 or 5x5, to test more intricate cases but that
will increase test run times quite a lot, for example, it takes about
15 minutes on my machine to run the tests on a grid of size 3x5 with a
release build.

The test can be run with a bigger grid by setting the
KWIN_TEST_REGION_GRID_SIZE environment variable, for example

    KWIN_TEST_REGION_GRID_SIZE=4x4 bin/testRegion

or if you want to run a specific test

    KWIN_TEST_REGION_GRID_SIZE=4x4 bin/testRegion united

Like with the Rect and RectF, the long term goal with the Region (and
RegionF) is to get them upstream in KF to help us make a stronger case
for changing upstream counter-parts in Qt or adopting our code.