wayland: Don't withdraw data offers when keyboard focus changes
When a dialog is closed, the keyboard focus doesn't move to the parent
window immediately. Instead, it jumps to the null window, and then from
the null window to the parent window.

The window activation bits need more massaging. In meanwhile, keep the
old data offers around. Not great, but it restores the old behavior.

BUG: 511509