improve support for tiled displays
Tiled displays are made of multiple parts, with each connector representing only
part of the actual physical display. To support them properly, this commit makes
- the backend provide information about each tile
- Workspace create only one single logical output for the physical display
- the Compositor adjust the scene views to the geometry of each BackendOutput

The only currently missing piece is ensuring atomic presentation of all
BackendOutputs that belong to one tiled display, but that can be fixed
separately later on.

BUG: 427545