Make Workspace::activateNextWindow() return type void
activateNextWindow() returns a boolean mainly because of the FocusIn
handler for the X11 root window.

If the focus policy is unreasonable, activateNextWindow() can call
focusToNull() for the FocusIn handler.

If focus updates are blocked, just ignore it and wait until
requestFocus() is called.

Not caring about the return type makes changing the activateNextWindow()
function easier.