Split Workspace::takeActivity()
Since the takeActivity() function is overloaded, it creates a dilemma if
a window needs to be both focused and raised and focus updates are
blocked.

If focus updates are blocked, ideally, we should return false. But what
about the raise action, can it still be performed, etc?

This function splits the takeActivity() function so the requestFocus()
implementation can be simpler and more sane.

Note that with this, the requestFocus() will return false if focus
updates are blocked or the window is a splash screen.