Fix focus not jumping immediately from closed dialog to the parent window
With the current implementation, the activateNextWindow() function
unsets the active window, then attempts to focus the next window.

This creates a problem when closing a dialog.

Due to the focus moving temporarily to a null window, all the current
data offers will be withdrawn. This can break the paste dialog in
dolphin.

With the new behavior, the active window will immediately jump from the
dialog to the parent window.

If focus updates are blocked, the null window will be focused. In long
term, we should look into killing the block_focus stuff to make focus
management less error prone.