autotests/integration: check kms state for drm autotests
The output layers could be "enabled" while the drm backend doesn't commit at all, or
puts it in the wrong z order, or actually renders while we wanted to test for direct scanout.
By fetching the drm state, we can test for the actual user-visible state (assuming
functional drivers of course).

Note that this commit makes the autotest require atomic modesetting to be supported
by the driver it's running on, even when running the legacy modesetting test. This is
needed because legacy modesetting doesn't really provide good ways to check the kernel
side state, so the test would be much less useful if we skipped those parts.