Use qreal instead of int to store tablet pad ring positions
While looking through the libinput API, we noticed that the
libinput_event_tablet_pad_get_ring_position actually returns a double -
not an integer. My own hardware doesn't emit anything this low, but it's
possible something does.

More importantly, the Wayland protocol accepts a double for ring
position events. So if there is hardware that allows shifting a ring in
quarter or half degrees, applications should be able to handle those
now.