plugins/trackmouse: Fix stuttering
With the output layer changes in 6.5, repaints for the next frame can't
be scheduled in the pre paint stage anymore.

On the other hand, the track mouse effect relied on that behavior. As a
short term solution, this change moves the rotation transform update
step to the post paint stage.

BUG: 512767