core/renderjournal: improve render time prediction algorithm more
The previous algorithm was very conservative, which in general is good at avoiding
dropped frames, but added a lot of latency when render times are just a little bit unstable.
In some cases, this added latency, and especially often unnecessary switches to triple
buffering, could make the system feel *less* responsive and animations less smooth than
if we'd just allow rare and extreme frames to be dropped.

The best example for a problematic situation is when the overview effect starts, which
has 1-3 frames of really terrible "render" times, but afterwards falls down to much lower
frame times. In that case, the algoithm made KWin render two frames in advance for tens
of seconds, which is quite bad.
Another common case is when new windows are opened, or the active window changes,
in which case KWin's render times can be increased by a few milliseconds just for that one
frame.

This new algorithm saves the render duration of the last 100 frames, and calculates the
average and standard deviation of that dataset.
With that information, it calculates a predicted render time, which is statistically very likely
to be higher than the next frame's actual render time, without taking individual spikes into
account too much.