plugins/screencast: Fix cursor metadata in output screencasts
Amends a49ab68672f40de4c56f0ba80d1b83737494a905.

The includesCursor() determines whether the cursor is inside the stream,
in general. It is used by both the metadata and the embedded code paths.

We must not synchronize it with the cursor view because it is for the
embedded code path.