x11: Remove Workspace::should_get_focus
At the moment, you need to be really really careful with choosing
whether to use the Workspace::activeWindow() or the
Workspace::mostRecentlyActivatedWindow(). As git history shows, we tend
to get it wrong.

With the recent active window management changes, the active window is
no longer updated upon receiving FocusIn events. Due to that, we can
start looking into getting rid of the should_get_focus queue and making
the Workspace APIs more intuitive.