input: Allow some devices to skip being routed through input methods
Some devices already provide text in a method that has already gone
through an input method.

We still want input methods available for key presses, but don't want
input from certain virtual devices to go through this.

There is not programatic way to handle this as we don't have appropriate
metadata available. As a simple stop-gap solution an environment
variable based on the device name allows vendors or users to create a
solution.

An example being the steam on screen keyboard that uses xtest events
which can now be guarded with
KWIN_DEVICE_SKIPS_INPUT_METHOD=steam\ eis\ keyboard