xdgshellwindow: allow forcing "no border" to false for clients that don't support SSD
Some applications don't do a good job with their client side decorations, so this allows the
user to force it on those apps anyways.

This may result in pretty bad visuals with some applications, so this keeps the functionality
restricted to window rules and does not expose it in window actions or plasma window
management protocol.