workspace: make dpms handling more correct
The previous code assumed that outputs would be immediately turned off in
`Workspace::applyOutputConfiguration` - but that is not correct (anymore) with
the drm backend, which does asynchronous modesets now.
To fix that mismatch, this adds a third dpms state to `BackendOutput`, so the
drm backend can specify if the output is merely turning off or if it actually
is off already.