plugins/zoom: Add strict version of MouseTrackingCentered
Some users find it difficult to interact with elements in the corners
or edges of their screen, while zoomed in. Other compositors address
this by allowing the zoom area to extend beyond the visible workspace,
keeping the mouse cursor in the center of the screen at all times.

This commit adds that functionality, in the form of a new strict version
of the existing Centered tracking mode.

The new mode is defined after Disabled, instead of being inserted
between existing modes, to prevent accidental config changes when
a user upgrades Plasma.

BUG: 513145
Signed-off-by: Ritchie Frodomar <alkalinethunder@gmail.com>