core/sessions: don't take ownership of an fd that Qt will close
QDBusUnixFileDescriptor closes the file descriptor in its destructor, if we
want to take ownership of it, we need to duplicate the file descriptor.
If we don't duplicate it, but directly take "ownership" of the file descriptor,
the following sequence of events can happen:

1. ~QDBusUnixFileDescriptor closes file descriptor A
2. some new fd B is allocated, which happens to have the same handle as fd A
3. ~FileDescriptor closes fd A, but it has the same handle as fd, so it actually
   closes fd B!
4. whenever fd B is used, we run into problems

To make such issues easier to debug in the future, this also adds debug logging
for when closing a file descriptor fails.

BUG: 513151