screenedge: Make touch target customizable from kconfig
This makes the touch threshold for gestures customizable from kconfig.
Currently it is hardcoded to 8 pixels, however this suitability of this
value depends on the display. On Plasma Mobile, for some phones (ex. Pixel 3a) it's almost impossible
to trigger the navigation panel with this value, but for others, it
works completely fine.