Allow setting a javascript callback to place windows
This changeset allows kwin scripts to register a "placement callback",
which, if set, will be called whenever kwin has to place a window. If
this callback is set and returns a non-null rectangle, this rectangle
will be considered when placing a window, at "fairly high priority"
(i.e. window rules will overrule, otherwise, it's very likely to
determine how the window is rendered). This also provides a hook to the
JS runtime to set the window's frameGeometry, border, etc. as it
provides access to the Kwin::Window pointer.

This allows us to implement custom, scripted window placement policies
without touching kwin's core.

Signed-off-by: Sebastian Kügler <sebas@kde.org>