Force complementing active/inactive opacity window rule
If only the inactive opacity is forced, the window will still use the
inactive opacity value in the active state.

In order to make the window rule make some sense, the opacity property
will need to be split in several sub-properties. On the other hand, only
the window rules are the ones that expect the opacity to be different
depending on the window state. Window actions assume there's only one
opacity. So, given that, this change makes kwin force an implicit
complementing opacity window rule (if no such a rule is provided), so
the opacity is reset back to `1.0`.