core: Reduce amount of XYWH <-> LTRB conversions in OutputTransform::map()
This gets rid of extra XYWH <-> LTRB conversions so the map() performs
only the necessary calculations.

The XYWH conversion happens when we ask for the width or the height of
the rect, and the LTRB conversion happens when we construct a Rect/RectF.

This shouldn't have a significance performance impact, it is mainly
to be more pedantic about doing unnecessary things.