kcms/desktop: improve renaming UX with keyboard
The current delegates do not support hitting Return on the on-demand
rename field to accept the changes, nor hitting Escape to cancel them.
While we're at it, we also block the user from moving the listview
currentitem with up/down keys while editing, as that is a pretty awkward
experience that requires awkward navigation to get back to the textfield
and allow pressing F2 to start renaming, which is the standard shortcut
for this action in applications (albeit not defined as a StandardKey,
so we can't match against that).