input: Unify side-ffects in TabletInputRedirection, expose last position
I created a similar-looking setPosition function to set some of the
needed side-effects such as confining the pointer. In the future I want
to allow unifying the last position of pointer and tablet devices, so we
need to keep track of that and allow "taking" the last known position of
either.