core/backendoutput: improve auto brightness curve adjustments
Previously, a sequence like
- set brightness to 40% at lux = 10
- set brightness to 20% at lux = 0
- set brightness to 35% at lux = 0

would result in a curve that sampled a percentage value of 20%, instead of the
expected 35%. This was because lux values in the curve were clamped to zero. In
addition, the calculation for high and low values was rather imprecise, so this
commit replaces it with a linear equation.

There was also an issue with the following sequence:
- set brightness to 40% at lux = 0
- set brightness to 20% at lux = 0

in which case a lux value of zero would result in a brightness of 40%, instead
of the expected 20%. To fix that, this makes the curve strictly monotonic.